Add files via upload
PIR Arduino
Using a PIR Sensor with Arduino
Circuit. You can connect PIR output to any digital pin. There is a jumper behind this module. If you move the jumper to L position, the sensor will 'toggle' (change state) whenever motion is detected. This is unlikely to be of much use in a practical applications
circuit with 
VCC-VCC
GND-GND
OUTPUT - ANY DIGITAL